delete logout link and use button instead, migration link only when token exists
