+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 // App.tsx (or any other component where you want to use the NavBarComponent)
 
-import * as React from 'react';
+import React, { useEffect } from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Route as Route } from 'react-router-dom';
 import { Routes as Routes } from 'react-router-dom';
@@ -10,38 +10,21 @@ import favicon from '/client/src/assets/favicon.svg';
 import logoImage from './assets/logo.png';
 import MigrationPage from './pages/MigrationPage';
 import SignUpPage from './pages/SignUpPage';
-
 import GitHubCallBack from './components/GitHubCallBack';
-// import GitHubLogin from './components/GitHubLogin';
 import DocumentationPage from './pages/DocumentationPage';
 import CommunityPage from './pages/CommunityPage';
 import FAQPage from './pages/FAQPage';
 import MainPage from './pages/MainPage';
-import LogOut from './components/LogOutButton';
-import UserBubble from './components/UserBubble'; // Might be good to implement for user/guest?
 import AddMigrationsPage from './pages/AddMigrationsPage';
 import UpdateMigrationsPage from './pages/UpdateMigrationsPage';
 
 const App: React.FC = () => {
-	const navLinks = [
-		{ path: '/documentation', label: 'Documentation' },
-		{ path: '/faq', label: 'FAQ' },
-		{ path: '/community', label: 'Community' },
-		{ path: '/migration', label: 'Migration' },
-		{ path: '/login', label: 'Login' },
-		{ path: '/signup', label: 'Signup' },
-		{ path: '/logout', label: 'Log Out'}
-	];
-
 	return (
 		<Router>
 			<div>
-				<NavBar links={navLinks} logo={logoImage} />
-				<UserBubble />
-				{/* <UsersList /> */}
-
+				<NavBar logo={logoImage} />
 				<Routes>
-					<Route path="/" element={<MainPage />} />
+					<Route path="/" element={<LoginPage />} />
 					<Route path="/login" element={<LoginPage />} />
 					<Route path="/updateMigrations" element={<UpdateMigrationsPage />} />
 					<Route path="/github/callback" element={<GitHubCallBack />} />
@@ -51,9 +34,6 @@ const App: React.FC = () => {
 					<Route path="/documentation" element={<DocumentationPage />} />
 					<Route path="/community" element={<CommunityPage />} />
 					<Route path="/faq" element={<FAQPage />} />
-					<Route path="/logout" element={<LogOut />} />
-					{/* Define more routes as necessary */}
-
 				</Routes>
 			</div>
 		</Router>
