+++ client/src/components/LogOutButton.tsx
@@ -1,11 +1,13 @@
-import * as React from 'react';
+// import * as React from 'react';
 
-function handleLogOut() {
-  sessionStorage.clear();
-}
+// function handleLogOut() {
+// 	sessionStorage.clear();
+// }
 
-const LogOut: React.FC = () => (
-  <button className="logOut" onClick={handleLogOut}>Log Out</button>
-);
+// const LogOut: React.FC = () => (
+// 	<button className="logOut" onClick={handleLogOut}>
+// 		Log Out
+// 	</button>
+// );
 
-export default LogOut;
+// export default LogOut;
