+++ client/src/components/NavBar.tsx
@@ -1,6 +1,6 @@
-import * as React from 'react'; // imports al;l exports from reacct pkg and bundles them as React
-import { Link } from 'react-router-dom'; // Link used to navigate btwn routes w/in app without page refresh
-
+import React, { useEffect } from 'react'; // imports al;l exports from reacct pkg and bundles them as React
+import { Link, useNavigate } from 'react-router-dom'; // Link used to navigate btwn routes w/in app without page refresh
+import UserBubble from './UserBubble'; // Might be good to implement for user/guest?
 interface NavLink {
 	// describes sha[pe of objects where each object has path and label type of string]
 	path: string;
@@ -8,13 +8,29 @@ interface NavLink {
 }
 
 interface NavBarProps {
-	// expects a link array and a logo string
-	links: NavLink[];
 	logo?: string; // Optional logo prop
 }
 
-const NavBar: React.FC<NavBarProps> = ({ links, logo }) => {
-	// declaring NavBar component and takes in NavBarProps
+const NavBar: React.FC<NavBarProps> = ({ logo }) => {
+	const navigate = useNavigate();
+	const navLinks: NavLink[] = [
+		{ path: '/documentation', label: 'Documentation' },
+		{ path: '/faq', label: 'FAQ' },
+		{ path: '/community', label: 'Community' },
+		{ path: '/login', label: 'Login' },
+		{ path: '/signup', label: 'Signup' },
+	];
+
+	if (sessionStorage.getItem('token')) {
+		navLinks.push({ path: '/migration', label: 'Migration' });
+	}
+
+	function handleLogOut() {
+		sessionStorage.clear();
+		navigate('/login');
+	}
+
+	console.log(navLinks);
 	return (
 		<nav>
 			<ul style={{ display: 'flex', alignItems: 'center', listStyle: 'none' }}>
@@ -23,13 +39,17 @@ const NavBar: React.FC<NavBarProps> = ({ links, logo }) => {
 						<img src={logo} alt="Logo" style={{ height: '50px' }} />
 					</li>
 				)}
-				{links.map((link, index) => (
+				{navLinks.map((link, index) => (
 					<li key={index} style={{ marginLeft: '20px' }}>
 						<Link to={link.path}>{link.label}</Link>
 					</li>
 				))}
 				<li style={{ marginLeft: 'auto' }}></li>
+				<button className="logOut" onClick={handleLogOut}>
+					Log Out
+				</button>
 			</ul>
+			<UserBubble />
 		</nav>
 	);
 };
